creating VatTu TronGoi API

+++ src/main/java/com/devmam/slmapiv2/services/JwtService.java
@@ -3,13 +3,14 @@ package com.devmam.slmapiv2.services;
 import com.nimbusds.jwt.JWTClaimsSet;
 
 import java.util.Set;
+import java.util.UUID;
 
 public interface JwtService {
     String buildScope(Set<String> roles);
 
     String generateToken(String id, String email, Set<String> roles, String userAgent);
 
-    Long getUserId(String token);
+    Integer getUserId(String token);
 
     String getKeycloakUserIdFromToken(String token);
